no message
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@38 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/tests/cv/src/adistancetransform.cpp
@@ -171,7 +171,7 @@ int test_dt(void* arg)
             euclid = FLT_MAX;
             for(ip = 0; ip < npoints; ip++)
             {
-                distance = (float)sqrt((xi-x[ip])*(xi-x[ip])+(yi-y[ip])*(yi-y[ip]));
+                distance = (float)sqrt((double)(xi-x[ip])*(xi-x[ip])+(yi-y[ip])*(yi-y[ip]));
                 if(distance < euclid)
                 {
                     euclid = distance;
